FELIX-4888 : ServletHandler's are not sorted by longest matching path. DTO handling (WiP)
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk/http/base@1680059 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/felix/http/base/internal/registry/FilterRegistry.java
@@ -53,7 +53,7 @@ public final class FilterRegistry
             return handler.getFilterInfo();
         }
         @Override
-        public long getReason() {
+        public int getReason() {
             return result;
         }
     }
